python/requirements.txt
simple-salesforce==1.12.9
python-dotenv>=1.0.0
